third_party/blink/web_tests/virtual/threaded-no-composited-antialiasing/animations/img-element-transform.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<img id="target" src="../../../animations/resources/green-100.png">
<script>
promise_test(function() {
  var animation = target.animate({transform: ['rotate(0deg)', 'rotate(180deg)']}, 100000);
  return animation.ready.then(() => {
    assert_true(internals.isCompositedAnimation(animation));
  });
}, '<img> elements should run compositor animations');
</script>
